+++ src/includes/class-asana-interface.php
@@ -32,11 +32,21 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		/**
 		 * The ?opt_fields csv for Asana API requests.
 		 *
-		 * @since 1.0.0
+		 * @since 3.1.0
 		 *
 		 * @var string TASK_OPT_FIELDS
 		 */
 		public const TASK_OPT_FIELDS = 'name,completed,notes,due_on,assignee,workspace,tags';
+
+		/**
+		 * The collection of all site tasks which are accessible by the currently
+		 * authenticated user, mapped by Asana task GID.
+		 *
+		 * @since 3.1.0
+		 *
+		 * @var \stdClass[] $all_site_tasks;
+		 */
+		public static $all_site_tasks;
 
 		/**
 		 * The currently authenticated WordPress user's ID.
@@ -144,6 +154,10 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		 * or requests could fail due to request limits or server issues.
 		 */
 		private static function maybe_load_client( int $user_id = 0 ) : \Asana\Client {
+
+			if ( 0 === $user_id ) {
+				$user_id = get_current_user_id();
+			}
 
 			$asana_personal_access_token = Options::get( Options::ASANA_PAT, $user_id );
 			if (
@@ -564,10 +578,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		 * @throws \Exception The Asana client may not be authenticated or the API
 		 * request may fail. Additional custom exceptions are:
 		 * * 400: Invalid task gid - The provided task gid is invalid.
-		 * * 410: Unpinned task - The API returned 404, so the task was unpinned.
-		 * * 410: Unpinned Foreign Task - The task does not belong to the assigned
-		 * workspace or have the site tag, so it was unpinned.
-		 * * 0: Failed to get task data - This is presumably unreachable.
+		 * * 410: Invalid task - The task is no longer available or relevent.
 		 */
 		public static function maybe_get_task_data( string $task_gid, string $opt_fields_deprecated = '', int $post_id = 0 ) : \stdClass {
 
@@ -581,78 +592,37 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 
 			$task_gid = Options::sanitize( 'gid', $task_gid );
 			if ( empty( $task_gid ) ) {
-				throw new \Exception( 'Invalid task gid', 400 );
+				throw new \Exception( 'Invalid task gid.', 400 );
 			}
 
-			// Get from cache.
-			$transient_key = Cache_Manager::get_cache_key( "task_{$task_gid}" );
-			$transient = get_transient( $transient_key );
-			if ( false !== $transient ) {
-				return $transient;
+			// If already in memory.
+			$all_site_tasks = self::maybe_get_all_site_tasks();
+			if ( ! empty( $all_site_tasks[ $task_gid ] ) ) {
+				return $all_site_tasks[ $task_gid ];
 			}
 
-			try {
-
-				$asana = self::get_client();
-				$task = $asana->tasks->findById( $task_gid, [ 'opt_fields' => self::TASK_OPT_FIELDS ] );
-
-				if (
-					isset( $task->workspace->gid )
-					&& $task->workspace->gid != Options::get( Options::ASANA_WORKSPACE_GID )
-					&& $post_id > 0
-				) {
-					if ( '' != $task_gid && Options::delete( Options::PINNED_TASK_GID, $post_id, $task_gid ) ) {
-						error_log( "Unpinned foreign task from post $post_id." );
-						throw new \Exception( 'Unpinned Foreign Task', 410 );
-					}
-				}
-
-				if (
-					isset( $task->tags )
-					&& is_array( $task->tags )
-					&& ! self::has_tag( $task, Options::get( Options::ASANA_TAG_GID ) )
-					&& $post_id > 0
-				) {
-					if ( '' != $task_gid && Options::delete( Options::PINNED_TASK_GID, $post_id, $task_gid ) ) {
-						error_log( "Unpinned task missing site tag from post $post_id." );
-						throw new \Exception( 'Unpinned Foreign Task', 410 );
-					}
-				}
-
-				set_transient( $transient_key, $task, Cache_Manager::get_transient_lifespan() );
-				return $task;
-			} catch ( \Exception $e ) {
-
-				$error_code = $e->getCode();
-				if (
-					0 === $error_code
-					&& isset( $e->status )
-					&& $e->status > 0
-				) {
-					$error_code = $e->status;
-				}
-
-				$error_msg = $e->getMessage();
-
-				if (
-					404 == $error_code
-					&& $post_id > 0
-				) {
-					if ( '' != $task_gid && Options::delete( Options::PINNED_TASK_GID, $post_id, $task_gid ) ) {
-						error_log( "Unpinned [404: Not Found] task from post $post_id." );
-						throw new \Exception( 'Unpinned Task', 410 );
-					}
-				} elseif (
-					403 != $error_code
-					&& 410 != $error_code
-				) {
-					error_log( "Failed to fetch task data, error $error_code: $error_msg" );
-				}
-
-				throw $e;
+			// Force cache update to ensure task is actually invalid.
+			/*
+			** Note that this is inefficient for invalidating many removed
+			** tasks since the ENTIRE cache object is being reloaded for every
+			** single invalid task. Use a bulk getter instead so this situation can
+			** be handled efficiently with just one forced update.
+			*/
+			$all_site_tasks = self::maybe_get_all_site_tasks( '', true );
+			if ( ! empty( $all_site_tasks[ $task_gid ] ) ) {
+				return $all_site_tasks[ $task_gid ];
 			}
 
-			throw new \Exception( 'Failed to get task data.', 0 );
+			// Task is no longer valid, handle accordingly.
+			if ( $post_id > 0 ) {
+				if ( Options::delete( Options::PINNED_TASK_GID, $post_id, $task_gid ) ) {
+					$err_msg = "Unpinned invalid task {$task_gid} from post {$post_id}.";
+					error_log( $err_msg );
+					throw new \Exception( $err_msg, 410 );
+				}
+			}
+
+			throw new \Exception( "Invalid task {$task_gid}.", 410 );
 		}
 
 		/**
@@ -703,16 +673,19 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 		/**
 		 * Attempts to retrieve task data for all site tasks.
 		 *
+		 * @since 3.1.0 Added optional param $force_update.
 		 * @since 3.1.0 Marked $opt_fields param as deprecated.
 		 * @since 1.0.0
 		 *
 		 * @param string $opt_fields_deprecated Deprecated.
+		 * @param bool $force_update Optional. If the cache should be invalidated
+		 * regardless of its expiration time. Default false.
 		 * @return \stdClass[] Task data objects.
 		 *
 		 * @throws \Exception Authentication may fail when first loading the client
 		 * or requests could fail due to request limits or server issues.
 		 */
-		public static function maybe_get_all_site_tasks( string $opt_fields_deprecated = '' ) : array {
+		public static function maybe_get_all_site_tasks( string $opt_fields_deprecated = '', bool $force_update = false ) : array {
 
 			if ( ! empty( $opt_fields_deprecated ) ) {
 				_deprecated_argument(
@@ -722,13 +695,20 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 				);
 			}
 
-			$transient_key = Cache_Manager::get_cache_key( 'all_site_tasks' );
-			$transient = get_transient( $transient_key );
-			if ( false !== $transient ) {
-				return $transient;
-			}
-
+			// Load client to ensure current user is authenticated and set.
 			$asana = self::get_client();
+
+			// Get all site tasks for the authenticated user.
+			$transient_key = Cache_Manager::get_cache_key( 'all_site_tasks_' . self::$wp_user_id );
+
+			// Use cached data if not forcing update.
+			if ( false === $force_update ) {
+				$transient = get_transient( $transient_key );
+				if (false !== $transient ) {
+					self::$all_site_tasks = $transient;
+					return $transient;
+				}
+			}
 
 			$tasks = [];
 
@@ -753,9 +733,10 @@ if ( ! class_exists( __NAMESPACE__ . '\Asana_Interface' ) ) {
 			$site_tasks = $asana->tasks->findByTag( $site_tag_gid, $params, $options );
 			$all_tasks = [];
 			foreach ( $site_tasks as $task ) {
-				$all_tasks[] = $task;
+				$all_tasks[ $task->gid ] = $task;
 			}
 
+			self::$all_site_tasks = $all_tasks;
 			set_transient( $transient_key, $all_tasks, Cache_Manager::get_transient_lifespan() );
 			return $all_tasks;
 		}