+++ changelog.md
@@ -6,7 +6,12 @@
 
 #### Changed
 
+- Implemented caching for all site tasks per authenticated user, vastly reducing API calls and improving performance.
 - Tasks dashboard widget has been converted to ReactJS, offering a better experience.
+
+#### Fixed
+
+- `Asana_Interface::$wp_user_id` would not be properly set to the current user's ID when loading the Asana client where user ID is `0`, the default value.
 
 ### 3.0.1 - 2021-11-09